requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.2
django-admin-autocomplete-filter==0.7.1
django-constance==2.9.1
django-environ==0.10.0
django-picklefield==3.1
django-mptt==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
django_filter==23.2
drf-yasg==1.21.5
dj-rest-auth==3.0.0
factory_boy==3.2.1
gunicorn==20.1.0
openpyxl==3.1.2
packaging==23.0
pandas==2.0.0
pathspec==0.11.1
platformdirs==3.2.0
psycopg2-binary==2.9.6
sqlparse==0.4.3
tomli==2.0.1
